+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 ### Fixed
 
 - Fixed issue where version would not be correctly processed if they only had a major and minor version, e.g. `1.2` or `v1.2`.
+- In case `CPM_USE_NAMED_CACHE_DIRECTORIES` is set, clearing the cache for a specific package, should now also remove the directory with the hash as name.
 
 ### Notes